Update: GerarToken() movido para ConfigurarToken.cs
