--- backend/ImaPay/Helpers/TokenSettings.cs
@@ -1,6 +1,0 @@
-﻿namespace ImaPay.Helpers;
-
-public class TokenSettings
-{
-    public static string SecretKey = "@&&N3>+Dl}'HWY4t?/I2gk}TU'XoSxj[Fy1^UX+;CJpd$eao[%kAO~m%ba9&oG:<(5cRQlq$q@Lz2QL$%'.o:jij<%,y8axn*A3?IFty*i4<;K@M{_I#NjA2%{4DXLLNTLOw*5c(Z]}5|h%Ri,V_k>YeFSLr26}A6%tpBz~G@E&E')@0:aI0Lpg=OCPrm7PfR5nA,??r5M8&s*7i3E@hZ+&H@cj~*x/VpwoMugy[DE_9.N*Xee5*o?0!6~V^>a_";
-}
